+++ include/fiction/utils/layout_utils.hpp
@@ -16,6 +16,7 @@
 #include <algorithm>
 #include <cassert>
 #include <cstdint>
+#include <cstdio>
 #include <limits>
 #include <random>
 #include <type_traits>
@@ -271,7 +272,12 @@ Lyt normalize_layout_coordinates(const Lyt& lyt) noexcept
 
     Lyt lyt_new{};
 
-    lyt_new.resize({lyt.x() - x_offset, lyt.y() - y_offset, lyt.z()});
+    assert(lyt.x() - x_offset >= 0 && "x_offset is too large");
+    assert(lyt.y() - y_offset >= 0 && "y_offset is too large");
+
+    lyt_new.resize(
+        {static_cast<std::size_t>(lyt.x() - x_offset), static_cast<std::size_t>(lyt.y() - y_offset), lyt.z()});
+
     lyt_new.set_layout_name(lyt.get_layout_name());
     lyt_new.set_tile_size_x(lyt.get_tile_size_x());
     lyt_new.set_tile_size_y(lyt.get_tile_size_y());
@@ -686,6 +692,110 @@ template <typename CoordinateType>
     }
 }
 #pragma GCC diagnostic pop
+
+/**
+ * This function checks whether the given layouts `first_lyt` and `second_lyt` are identical by comparing various
+ * properties such as the number of cells, the types of cells, defects (if applicable), and charge states (if
+ * applicable). The comparison is done in a detailed manner depending on the specific layout type.
+ *
+ * @Note The aspect ratios of the cell-level layouts are not compared.
+ *
+ * @tparam Lyt The layout type. Must be a cell-level layout.
+ * @param first_lyt The first layout to compare.
+ * @param second_lyt The second layout to compare.
+ * @return `true` if the layouts are identical, `false` otherwise.
+ */
+template <typename Lyt>
+[[nodiscard]] inline bool are_cell_layouts_identical(const Lyt& first_lyt, const Lyt& second_lyt) noexcept
+{
+    static_assert(is_cell_level_layout_v<Lyt>, "Lyt is not a cell-level layout");
+
+    if (first_lyt.num_cells() != second_lyt.num_cells())
+    {
+        return false;
+    }
+
+    bool different_cells = false;
+
+    first_lyt.foreach_cell(
+        [&first_lyt, &second_lyt, &different_cells](const auto& c)
+        {
+            if (first_lyt.get_cell_type(c) != second_lyt.get_cell_type(c))
+            {
+                different_cells = true;
+                return false;  // abort
+            }
+            return true;  // keep looping
+        });
+
+    if (different_cells)
+    {
+        return false;
+    }
+
+    if constexpr (is_sidb_defect_surface_v<Lyt>)
+    {
+        if (second_lyt.num_defects() != first_lyt.num_defects())
+        {
+            return false;
+        }
+
+        bool different_defects = false;
+
+        first_lyt.foreach_sidb_defect(
+            [&first_lyt, &second_lyt, &different_defects](const auto& defect_old)
+            {
+                if (first_lyt.get_sidb_defect(defect_old.first) != second_lyt.get_sidb_defect(defect_old.first))
+                {
+                    different_defects = true;
+                    return false;  // abort
+                }
+                return true;  // keep looping
+            });
+
+        if (different_defects)
+        {
+            return false;
+        }
+    }
+
+    if constexpr (is_charge_distribution_surface_v<Lyt>)
+    {
+        if (second_lyt.num_neutral_sidbs() != first_lyt.num_neutral_sidbs())
+        {
+            return false;
+        }
+
+        if (second_lyt.num_negative_sidbs() != first_lyt.num_negative_sidbs())
+        {
+            return false;
+        }
+
+        if (second_lyt.num_positive_sidbs() != first_lyt.num_positive_sidbs())
+        {
+            return false;
+        }
+
+        bool different_charge_state = false;
+        first_lyt.foreach_cell(
+            [&different_charge_state, &first_lyt, &second_lyt](const auto& c)
+            {
+                if (first_lyt.get_charge_state(c) != second_lyt.get_charge_state(c))
+                {
+                    different_charge_state = true;
+                    return false;  // abort
+                }
+                return true;  // keep looping
+            });
+
+        if (different_charge_state)
+        {
+            return false;
+        }
+    }
+
+    return true;
+}
 
 }  // namespace fiction